clippy: tracing: allow needless_borrow on `record`
tracing-attributes 0.1.23 ships with a smarter `Span::record` that
allows owned values [^1], which makes

This means that previously invalid code such as:

    span.record("num_updates", num_items);

Is now valid, which makes clippy throw a warning for needless_borrow

Fixing the lint, however, means that foca would require a strict version
check for tracing-attributes (otherwise anything pulling 0.1.22 or
bellow would not compile anymore).

I don't want a strict dependency on tracing: it's tangential do foca and
I'd like users to benefit from the improvements that might be shipped
along, so I fix the lint warning by ignoring it instead.

Once we get a (at least minor) version bump on `tracing`, this can be
removed and updated to skip the borrow.

[^1]: https://github.com/tokio-rs/tracing/pull/2212