Periodic gossip: only send when there are updates
Technically, a gossip with no updates is still useful: it tells the
receiver that the sender is active so it helps with discovery and,
now, with the member gaining awareness that the cluster thinks its
dead.

It is, however, too chatty. During normal operations the updates buffer
is mostly empty; If we start sending messages with no updates too often
it won't be much different from a traditional gossip protocol with a
heartbeat.

So, since pings and announces already cover the case of talking to
the cluster unconditionally, it makes sense to make periodic gossiping
"smarter"