bugfix: Feed messages can come with trailing junk
When crafting a Message::Feed payload, we were assuming that codecs
wouldn't leave ANY invalid data in the buffer when an error
occurred

This is very likely to happen when the messages being
sent would reach Config::max_packet_size (i.e.: large clusters,
large identities) and the usd codec is not careful checking if
the whole Member can fit the buffer before writing anything to
it.

Whilst annoying, the errors this led to were harmless noise
unless there's a chance a codec would interpret the trailing
bytes as part of the previous entry in the buffer- quite
unlikely I think?

Before the fix, the added test fails with:

> ---- tests::feed_does_not_contain_trailing_jumk stdout ----
> thread 'tests::feed_does_not_contain_trailing_jumk' panicked at 'assertion failed: `(left == right)`
>   left: `Ok(())`,
>  right: `Err(CustomBroadcast(Broadcasts disabled))`', src/lib.rs:3634:9