React to TurnUndead messages from known down members
In the case of a partition, both sides will consider each other down.
Without this patch, when this situation happened the nodes would
engage in spamming each other with TurnUndead messages and never
actually reacting to it

This patch ensures that foca attempts to renew its own identity
before notifying the sender. In the worst case, both sides of the
partition will end up renewing their identities.