example: use a single task to handle timer events
before this patch, if the async runtime lagged for whatever reason
(system pause, very high load...), there was a chance the agent would
receive timer events out of order

this ensures the events will always be received in the order they
were supposed to happen. it's a bit more code than just YOLOing with
spawn() + sleep(), but a lot more robust and hopefully still easy
enough to grok for an example.