Prepare for v1.0.0
Experience tells me that as soon as I release this something will pop up
so here I am, goading the Fates.

The API hasn't changed in a long while, so it isn't much of a concern.
Not every dependency is stable, but they rarely change. A paced major
version bump simply to update deps is perfectly fine